% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_batch_write}
\alias{clouddirectory_batch_write}
\title{Performs all the write operations in a batch}
\usage{
clouddirectory_batch_write(DirectoryArn, Operations)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory.
For more information, see arns.}

\item{Operations}{[required] A list of operations that are part of the batch.}
}
\description{
Performs all the write operations in a batch. Either all the operations succeed or none.

See \url{https://paws-r.github.io/docs/clouddirectory/batch_write.html} for full documentation.
}
\keyword{internal}
