% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_create_schema}
\alias{clouddirectory_create_schema}
\title{Creates a new schema in a development state}
\usage{
clouddirectory_create_schema(Name)
}
\arguments{
\item{Name}{[required] The name that is associated with the schema. This is unique to each
account and in each region.}
}
\description{
Creates a new schema in a development state. A schema can exist in three phases:

See \url{https://paws-r.github.io/docs/clouddirectory/create_schema.html} for full documentation.
}
\keyword{internal}
