% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_attached_indices}
\alias{clouddirectory_list_attached_indices}
\title{Lists indices attached to the specified object}
\usage{
clouddirectory_list_attached_indices(
  DirectoryArn,
  TargetReference,
  NextToken = NULL,
  MaxResults = NULL,
  ConsistencyLevel = NULL
)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory.}

\item{TargetReference}{[required] A reference to the object that has indices attached.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}

\item{ConsistencyLevel}{The consistency level to use for this operation.}
}
\description{
Lists indices attached to the specified object.

See \url{https://paws-r.github.io/docs/clouddirectory/list_attached_indices.html} for full documentation.
}
\keyword{internal}
