% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_directories}
\alias{clouddirectory_list_directories}
\title{Lists directories created within an account}
\usage{
clouddirectory_list_directories(
  NextToken = NULL,
  MaxResults = NULL,
  state = NULL
)
}
\arguments{
\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}

\item{state}{The state of the directories in the list. Can be either Enabled,
Disabled, or Deleted.}
}
\description{
Lists directories created within an account.

See \url{https://paws-r.github.io/docs/clouddirectory/list_directories.html} for full documentation.
}
\keyword{internal}
