% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_add_custom_attributes}
\alias{cognitoidentityprovider_add_custom_attributes}
\title{Adds additional user attributes to the user pool schema}
\usage{
cognitoidentityprovider_add_custom_attributes(UserPoolId, CustomAttributes)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to add custom
attributes.}

\item{CustomAttributes}{[required] An array of custom attributes, such as Mutable and Name.}
}
\description{
Adds additional user attributes to the user pool schema.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/add_custom_attributes.html} for full documentation.
}
\keyword{internal}
