% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_enable_user}
\alias{cognitoidentityprovider_admin_enable_user}
\title{Enables the specified user as an administrator}
\usage{
cognitoidentityprovider_admin_enable_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to enable the user.}

\item{Username}{[required] The user name of the user you want to enable.}
}
\description{
Enables the specified user as an administrator. Works on any user.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/admin_enable_user.html} for full documentation.
}
\keyword{internal}
