% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_forget_device}
\alias{cognitoidentityprovider_admin_forget_device}
\title{Forgets the device, as an administrator}
\usage{
cognitoidentityprovider_admin_forget_device(UserPoolId, Username, DeviceKey)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{Username}{[required] The user name.}

\item{DeviceKey}{[required] The device key.}
}
\description{
Forgets the device, as an administrator.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/admin_forget_device.html} for full documentation.
}
\keyword{internal}
