% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_user_pool_domain}
\alias{cognitoidentityprovider_describe_user_pool_domain}
\title{Gets information about a domain}
\usage{
cognitoidentityprovider_describe_user_pool_domain(Domain)
}
\arguments{
\item{Domain}{[required] The domain string. For custom domains, this is the fully-qualified
domain name, such as \code{auth.example.com}. For Amazon Cognito prefix
domains, this is the prefix alone, such as \code{auth}.}
}
\description{
Gets information about a domain.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/describe_user_pool_domain.html} for full documentation.
}
\keyword{internal}
