% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_remove_ip_routes}
\alias{directoryservice_remove_ip_routes}
\title{Removes IP address blocks from a directory}
\usage{
directoryservice_remove_ip_routes(DirectoryId, CidrIps)
}
\arguments{
\item{DirectoryId}{[required] Identifier (ID) of the directory from which you want to remove the IP
addresses.}

\item{CidrIps}{[required] IP address blocks that you want to remove.}
}
\description{
Removes IP address blocks from a directory.

See \url{https://paws-r.github.io/docs/directoryservice/remove_ip_routes.html} for full documentation.
}
\keyword{internal}
