% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_settings}
\alias{directoryservice_update_settings}
\title{Updates the configurable settings for the specified directory}
\usage{
directoryservice_update_settings(DirectoryId, Settings)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to update settings.}

\item{Settings}{[required] The list of Setting objects.}
}
\description{
Updates the configurable settings for the specified directory.

See \url{https://paws-r.github.io/docs/directoryservice/update_settings.html} for full documentation.
}
\keyword{internal}
