% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_trust}
\alias{directoryservice_update_trust}
\title{Updates the trust that has been set up between your Managed Microsoft AD
directory and an self-managed Active Directory}
\usage{
directoryservice_update_trust(TrustId, SelectiveAuth = NULL)
}
\arguments{
\item{TrustId}{[required] Identifier of the trust relationship.}

\item{SelectiveAuth}{Updates selective authentication for the trust.}
}
\description{
Updates the trust that has been set up between your Managed Microsoft AD directory and an self-managed Active Directory.

See \url{https://paws-r.github.io/docs/directoryservice/update_trust.html} for full documentation.
}
\keyword{internal}
