% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_login_profile}
\alias{iam_delete_login_profile}
\title{Deletes the password for the specified IAM user, which terminates the
user's ability to access Amazon Web Services services through the Amazon
Web Services Management Console}
\usage{
iam_delete_login_profile(UserName)
}
\arguments{
\item{UserName}{[required] The name of the user whose password you want to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the password for the specified IAM user, which terminates the user's ability to access Amazon Web Services services through the Amazon Web Services Management Console.

See \url{https://paws-r.github.io/docs/iam/delete_login_profile.html} for full documentation.
}
\keyword{internal}
