% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_role}
\alias{iam_delete_role}
\title{Deletes the specified role}
\usage{
iam_delete_role(RoleName)
}
\arguments{
\item{RoleName}{[required] The name of the role to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the specified role. The role must not have any policies attached. For more information about roles, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles.html}{Working with roles}.

See \url{https://paws-r.github.io/docs/iam/delete_role.html} for full documentation.
}
\keyword{internal}
