% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_delete_custom_key_store}
\alias{kms_delete_custom_key_store}
\title{Deletes a custom key store}
\usage{
kms_delete_custom_key_store(CustomKeyStoreId)
}
\arguments{
\item{CustomKeyStoreId}{[required] Enter the ID of the custom key store you want to delete. To find the ID
of a custom key store, use the
\code{\link[=kms_describe_custom_key_stores]{describe_custom_key_stores}}
operation.}
}
\description{
Deletes a \href{https://docs.aws.amazon.com/kms/latest/developerguide/custom-key-store-overview.html}{custom key store}. This operation does not delete the CloudHSM cluster that is associated with the custom key store, or affect any users or keys in the cluster.

See \url{https://paws-r.github.io/docs/kms/delete_custom_key_store.html} for full documentation.
}
\keyword{internal}
