% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_security_hub}
\alias{securityhub_disable_security_hub}
\title{Disables Security Hub in your account only in the current Region}
\usage{
securityhub_disable_security_hub()
}
\description{
Disables Security Hub in your account only in the current Region. To disable Security Hub in all Regions, you must submit one request per Region where you have enabled Security Hub.

See \url{https://paws-r.github.io/docs/securityhub/disable_security_hub.html} for full documentation.
}
\keyword{internal}
