% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_master_account}
\alias{securityhub_get_master_account}
\title{This method is deprecated}
\usage{
securityhub_get_master_account()
}
\description{
This method is deprecated. Instead, use \code{\link[=securityhub_get_administrator_account]{get_administrator_account}}.

See \url{https://paws-r.github.io/docs/securityhub/get_master_account.html} for full documentation.
}
\keyword{internal}
