% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_security_hub_configuration}
\alias{securityhub_update_security_hub_configuration}
\title{Updates configuration options for Security Hub}
\usage{
securityhub_update_security_hub_configuration(AutoEnableControls = NULL)
}
\arguments{
\item{AutoEnableControls}{Whether to automatically enable new controls when they are added to
standards that are enabled.

By default, this is set to \code{true}, and new controls are enabled
automatically. To not automatically enable new controls, set this to
\code{false}.}
}
\description{
Updates configuration options for Security Hub.

See \url{https://paws-r.github.io/docs/securityhub/update_security_hub_configuration.html} for full documentation.
}
\keyword{internal}
